refactoring

+++ src/com/kopytko/Main.java
@@ -11,10 +11,10 @@ public class Main {
         byte b1 = 5;
         byte b2 = -4;
         System.out.println("b1 = " + b1 + ", b2 = " + b2);
-        System.out.println("Addition: " + (byte)(b1+b2));
-        System.out.println("Subtraction: " + (byte)(b1-b2) + " or " + (byte)(b2-b1) );
-        System.out.println("Multiplication: " + (byte)(b1*b2));
-        System.out.println("Division: " + (byte)(b1/b2) + " or " + (byte)(b1/b2));
+        System.out.println("Addition: " + (b1 + b2));
+        System.out.println("Subtraction: " + (b1 - b2) + " or " + (b2 - b1));
+        System.out.println("Multiplication: " + (byte) (b1 * b2));
+        System.out.println("Division: " + (byte) (b1 / b2) + " or " + (byte) (b1 / b2));
         System.out.println("----------------------");
         System.out.println();
 
@@ -23,10 +23,10 @@ public class Main {
         short s1 = -700;
         short s2 = 400;
         System.out.println("s1 = " + s1 + ", s2 = " + s2);
-        System.out.println("Addition: " + (short)(s1+s2));
-        System.out.println("Subtraction: " + (short)(s1-s2) + " or " + (short)(s2-s1) );
-        System.out.println("Multiplication: " + (short)(s1*s2));
-        System.out.println("Division: " + (short)(s1/s2) + " or " + (short)(s2/s1));
+        System.out.println("Addition: " + (s1 + s2));
+        System.out.println("Subtraction: " + (s1 - s2) + " or " + (s2 - s1));
+        System.out.println("Multiplication: " + (short) (s1 * s2));
+        System.out.println("Division: " + (short) (s1 / s2) + " or " + (short) (s2 / s1));
         System.out.println("----------------------------");
         System.out.println();
 
@@ -35,10 +35,10 @@ public class Main {
         long l1 = -145007L;
         long l2 = 17628L;
         System.out.println("l1 = " + l1 + ", l2 = " + l2);
-        System.out.println("Addition: " + (l1+l2));
-        System.out.println("Subtraction: " + (l1-l2) + " or " + (l2-l1) );
-        System.out.println("Multiplication: " + (l1*l2));
-        System.out.println("Division: " + (l1/l2) + " or " + (l2/l1));
+        System.out.println("Addition: " + (l1 + l2));
+        System.out.println("Subtraction: " + (l1 - l2) + " or " + (l2 - l1));
+        System.out.println("Multiplication: " + (l1 * l2));
+        System.out.println("Division: " + (l1 / l2) + " or " + (l2 / l1));
         System.out.println("---------------------------------");
         System.out.println();
 
@@ -47,10 +47,10 @@ public class Main {
         float f1 = 485689.00004f;
         float f2 = 63593.0562f;
         System.out.println("f1 = " + f1 + ", f2 = " + f2);
-        System.out.println("Addition: " + (f1+f2));
-        System.out.println("Subtraction: " + (f1-f2) + " or " + (f2-f1) );
-        System.out.println("Multiplication: " + (f1*f2));
-        System.out.println("Division: " + (f1/f2) + " or " + (f2/f1));
+        System.out.println("Addition: " + (f1 + f2));
+        System.out.println("Subtraction: " + (f1 - f2) + " or " + (f2 - f1));
+        System.out.println("Multiplication: " + (f1 * f2));
+        System.out.println("Division: " + (f1 / f2) + " or " + (f2 / f1));
         System.out.println("---------------------------------------");
         System.out.println();
 
@@ -59,17 +59,11 @@ public class Main {
         double d1 = -459485689.00004;
         double d2 = 63534893.0562;
         System.out.println("d1 = " + d1 + ", d2 = " + d2);
-        System.out.println("Addition: " + (d1+d2));
-        System.out.println("Subtraction: " + (d1-d2) + " or " + (d2-d1) );
-        System.out.println("Multiplication: " + (d1*d2));
-        System.out.println("Division: " + (d1/d2) + " or " + (d2/d1));
+        System.out.println("Addition: " + (d1 + d2));
+        System.out.println("Subtraction: " + (d1 - d2) + " or " + (d2 - d1));
+        System.out.println("Multiplication: " + (d1 * d2));
+        System.out.println("Division: " + (d1 / d2) + " or " + (d2 / d1));
         System.out.println("------------------------------------------------------------");
-
-
-
-
-
-
 
 
     }
